fix: MET-615 SPO's pagination was missing
