+++ src/components/StakingLifeCycle/SPOLifecycle/PoolUpdates/RecentPoolUpdates/PoolUpdatesList.test.tsx
@@ -9,10 +9,8 @@ describe("Pool Updates", () => {
   it("should render SPO PoollUpdatesList coponents", async () => {
     const setSelected = jest.fn();
     jest.spyOn(Router, "useParams").mockReturnValue({ useParams: "1" });
-    render(<PoollUpdatesList onSelect={setSelected} />);
+    render(<PoollUpdatesList onSelect={setSelected} setShowBackButton={jest.fn()} />);
     expect(screen.getByText("Recent Updates")).toBeInTheDocument();
-    const elm = screen.getByTestId("showing");
-    expect(elm).toHaveTextContent("Showing");
   });
 
   it("should render SPO table has 0 data", async () => {
@@ -22,9 +20,8 @@ describe("Pool Updates", () => {
         return [];
       }
     }));
-    render(<PoollUpdatesList onSelect={jest.fn()} />);
+    render(<PoollUpdatesList onSelect={jest.fn()} setShowBackButton={jest.fn()} />);
     expect(screen.getByText("Recent Updates")).toBeInTheDocument();
-    const elm = screen.getByTestId("showing");
-    expect(elm).toHaveTextContent("Showing 0 result");
+    expect(screen.getByText(/showing 0 result/i)).toBeInTheDocument();
   });
 });
